requirements.txt
pandas==1.5.3
numpy==1.23.5
seaborn==0.12.2
pip==23.1.2
scikit-learn==1.3.2
tensorflow==2.10.0
keras-tuner==1.1.0
streamlit==1.16.0
